fix language repository fix tests
