+++ app/src/main/java/ru/easycode/words504/languages/data/cache/LanguageKeyMapper.kt
@@ -1,0 +1,5 @@
+package ru.easycode.words504.languages.data.cache
+
+class LanguageKeyMapper: LanguageCache.Mapper<String> {
+    override fun map(key: String, name: String) = key
+}
