+++ app/src/main/java/ru/easycode/words504/languages/data/cloud/LanguageCloudCacheMapper.kt
@@ -2,6 +2,6 @@ package ru.easycode.words504.languages.data.cloud
 
 import ru.easycode.words504.languages.data.cache.LanguageCache
 
-class CloudCacheMapper : LanguageCloud.Mapper<LanguageCache> {
+class LanguageCloudCacheMapper : LanguageCloud.Mapper<LanguageCache> {
     override fun map(languageCode: String, name: String) = LanguageCache.Base(languageCode, name)
 }
