+++ app/src/main/java/ru/easycode/words504/languages/data/repository/LanguagesRepository.kt
@@ -6,17 +6,19 @@ import ru.easycode.words504.languages.data.cache.LanguagesCacheDataSource
 import ru.easycode.words504.languages.domain.LanguageDomain
 import ru.easycode.words504.languages.domain.LanguagesRepository
 
-class LanguagesSaveRepository(
+class LanguagesRepository(
     private val cacheDataSource: LanguagesCacheDataSource.Read,
     private val chosenLanguageCache: ChosenLanguageCache.Save,
-    private val domainMapper: LanguageCache.Mapper<LanguageDomain>
-) : LanguagesRepository.Languages, LanguagesRepository.Save {
+    private val domainMapper: LanguageCache.Mapper<LanguageDomain>,
+    private val keyMapper: LanguageCache.Mapper<String>
+) : LanguagesRepository.Languages {
 
     override suspend fun languages(): List<LanguageDomain> = cacheDataSource.read()
         .map { it.map(domainMapper) }
 
     override suspend fun save(languageCode: String) {
-        val cacheLanguage = cacheDataSource.read().find { it.key() == languageCode }
+        val cacheLanguage = cacheDataSource.read()
+            .find { it.map(keyMapper) == languageCode }
         cacheLanguage?.let {
             chosenLanguageCache.save(it)
         }
